+++ CHANGELOG.md
@@ -6,6 +6,9 @@ CHANGELOG versions 1.x
  * Add helper methods in base Command class ([#31](../../issues/31))
  * Update dependencies ([#33](../../issues/33))
  * Refactor MonologServiceProvider ([#35](../../issues/35))
+ * Add Monolog TimeExecutionProcessor processor ([#39](../../issues/39))
+ * Add Monolog Dumper util ([#40](../../issues/40))
+ * [git diff](../../compare/v1.1.4...v1.2.1)
 
 * v1.1 (2014-11-10)
 
@@ -14,8 +17,9 @@ CHANGELOG versions 1.x
  * Портирован `translateLevel` из `Silex` ([#18](../../issues/18))
  * Все подписчики на канал логирования помещены в свой реестр, по аналогии с `dbs` ([#17](../../issues/17))
  * Добавлена поддержка `ConsoleHandler` ([#16](../../issues/16))
- * [git diff](../../compare/v1.0...v1.1)
+ * [git diff](../../compare/v1.0.1...v1.1.4)
 
 * v1.0 (2014-08-30)
 
  * First stable release
+ * [git diff](../../compare/v0.6...v1.0.1)